requirements.txt
fastapi>=0.68.0,<0.69.0
uvicorn>=0.15.0,<0.16.0
sqlalchemy>=1.4.0,<1.5.0
pydantic>=1.8.0,<1.9.0
passlib>=1.7.4,<1.8.0
bcrypt>=3.2.0,<3.3.0
python-jose>=3.3.0,<3.4.0
python-multipart>=0.0.5,<0.0.6
alembic>=1.7.4,<1.8.0
psycopg2-binary>=2.9.1,<2.10.0
pytest>=6.2.5,<6.3.0
requests>=2.26.0,<2.27.0